pyproject.toml
[tool.poetry]
name = "duelpre"
version = "0.1.0"
description = ""
authors = ["ryouhama <hary0730cap@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
mysqlclient = "^2.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
